+++ src/alttxt/parser.py
@@ -1,9 +1,9 @@
 import json
 
-from alttxt.types_ import AggregateBy
-from alttxt.types_ import FileType
-from alttxt.types_ import SortBy
-from alttxt.types_ import SortVisibleBy
+from alttxt.types import AggregateBy
+from alttxt.types import FileType
+from alttxt.types import SortBy
+from alttxt.types import SortVisibleBy
 
 from alttxt.models import BookmarkedIntersectionModel
 from alttxt.models import DataModel
