+++ src/alttxt/types.py
