+++ src/alttxt/webapi.py
@@ -2,9 +2,9 @@ from fastapi import APIRouter
 from fastapi import FastAPI
 from pathlib import Path
 
-from alttxt.types_ import Granularity
-from alttxt.types_ import Level
-from alttxt.types_ import Orientation
+from alttxt.types import Granularity
+from alttxt.types import Level
+from alttxt.types import Orientation
 
 from alttxt.generic import Grammar
 from alttxt.generic import RawData
